+++ tls/src/rustls_uring.rs
@@ -12,7 +12,7 @@ pub use rustls_crate::*;
 
 use xitca_io::{
     bytes::{Buf, BytesMut},
-    io_uring::{AsyncBufRead, AsyncBufWrite, IoBuf, IoBufMut},
+    io_uring::{AsyncBufRead, AsyncBufWrite, BoundedBuf, BoundedBufMut},
 };
 
 use self::buf::WriteBuf;
@@ -61,17 +61,16 @@ impl<C, S> Session<C>
 where
     C: DerefMut + Deref<Target = ConnectionCommon<S>>,
 {
-    fn read_plain(&mut self, buf: &mut impl IoBufMut) -> io::Result<usize> {
-        io::Read::read(&mut self.session.reader(), io_ref_mut_slice(buf)).map(|n| {
+    fn read_plain(&mut self, buf: &mut impl BoundedBufMut) -> io::Result<usize> {
+        io::Read::read(&mut self.session.reader(), io_ref_mut_slice(buf)).inspect(|n| {
             // SAFETY
             // required by IoBufMut trait. when n bytes is write into buffer this method
             // must be called to advance the initialized part of it.
-            unsafe { buf.set_init(n) };
-            n
+            unsafe { buf.set_init(*n) };
         })
     }
 
-    fn write_plain(&mut self, buf: &impl IoBuf) -> io::Result<usize> {
+    fn write_plain(&mut self, buf: &impl BoundedBuf) -> io::Result<usize> {
         let writer = &mut self.session.writer();
         let n = io::Write::write(writer, io_ref_slice(buf))?;
         // keep this no op in case rustls change it's behavior.
@@ -110,10 +109,7 @@ where
             }
         }
 
-        let res = session.session.read_tls(&mut buf.as_ref()).map(|n| {
-            buf.advance(n);
-            n
-        });
+        let res = session.session.read_tls(&mut buf.as_ref()).inspect(|n| buf.advance(*n));
 
         session.read_buf.replace(buf);
 
@@ -229,7 +225,7 @@ where
 {
     async fn read<B>(&self, mut buf: B) -> (io::Result<usize>, B)
     where
-        B: IoBufMut,
+        B: BoundedBufMut,
     {
         let mut session = self.session.borrow_mut();
 
@@ -258,7 +254,7 @@ where
 {
     async fn write<B>(&self, buf: B) -> (io::Result<usize>, B)
     where
-        B: IoBuf,
+        B: BoundedBuf,
     {
         let mut session = self.session.borrow_mut();
 
@@ -308,13 +304,13 @@ where
     }
 }
 
-fn io_ref_slice(buf: &impl IoBuf) -> &[u8] {
+fn io_ref_slice(buf: &impl BoundedBuf) -> &[u8] {
     // SAFETY
     // have to trust IoBuf implementor to provide valid pointer and it's length.
     unsafe { slice::from_raw_parts(buf.stable_ptr(), buf.bytes_init()) }
 }
 
-fn io_ref_mut_slice(buf: &mut impl IoBufMut) -> &mut [u8] {
+fn io_ref_mut_slice(buf: &mut impl BoundedBufMut) -> &mut [u8] {
     // SAFETY
     // have to trust IoBufMut implementor to provide valid pointer and it's capacity.
     unsafe { slice::from_raw_parts_mut(buf.stable_mut_ptr(), buf.bytes_total()) }
@@ -343,13 +339,7 @@ mod buf {
             let (res, buf) = io.write(self.buf).await;
             self.buf = buf;
 
-            (
-                res.map(|n| {
-                    self.buf.advance(n);
-                    n
-                }),
-                self,
-            )
+            (res.inspect(|n| self.buf.advance(*n)), self)
         }
     }