+++ server/CHANGES.md
@@ -4,6 +4,7 @@
 - update `xitca-io` to `0.4.0`
 - update `xitca-service` to `0.2.0`
 - update `xitca-unsafe-collection` to `0.2.0`
+- update `tokio-uring` to `0.5.0`
 
 # 0.3.0
 ## Remove